TestLab Visitors: Move visitors in separate package
