+++ dmn-signavio/src/main/java/com/gs/dmn/signavio/testlab/visitor/EnhancerContext.java
@@ -10,9 +10,10 @@
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package com.gs.dmn.signavio.testlab;
+package com.gs.dmn.signavio.testlab.visitor;
 
 import com.gs.dmn.ast.TItemDefinition;
+import com.gs.dmn.signavio.testlab.TestLab;
 
 public class EnhancerContext {
     private final TestLab testLab;
