+++ dmn-signavio/src/main/java/com/gs/dmn/signavio/testlab/visitor/NopTestLabVisitor.java
@@ -10,8 +10,9 @@
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package com.gs.dmn.signavio.testlab;
+package com.gs.dmn.signavio.testlab.visitor;
 
+import com.gs.dmn.signavio.testlab.*;
 import com.gs.dmn.signavio.testlab.expression.*;
 
 public class NopTestLabVisitor<C> implements Visitor<TestLabElement, C> {
