+++ dmn-signavio/src/main/java/com/gs/dmn/signavio/testlab/visitor/TestLabEnhancer.java
@@ -10,10 +10,11 @@
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package com.gs.dmn.signavio.testlab;
+package com.gs.dmn.signavio.testlab.visitor;
 
 import com.gs.dmn.ast.TItemDefinition;
 import com.gs.dmn.runtime.DMNRuntimeException;
+import com.gs.dmn.signavio.testlab.*;
 import com.gs.dmn.signavio.testlab.expression.ComplexExpression;
 import com.gs.dmn.signavio.testlab.expression.Expression;
 import com.gs.dmn.signavio.testlab.expression.ListExpression;
